requirements.txt
chess==1.10.0
SpeechRecognition==3.10.1
pydub==0.25.1
python-dotenv==1.0.0
PyAudio==0.2.14
pygame==2.5.2
gTTS==2.4.0
transformers
torch
accelerate